convenience script for running on a robot
This update makes it easier to run the system on a robot.
The new convenience script start_web_server_and_robot_browser.sh
starts both the web server and the robot's browser on the robot.
When run in a terminal this script can be exited with Ctrl+C.
It can then be restarted.

Prior to running this new convenience script the appropriate
ROS launch file should be run. However, the ROS launch file
can be kept running while this script is exited and restarted.

This new convenience script calls start_robot_browser.js,
which uses puppeteer to launch the robot's browser and
log the robot in.

WARNING: start_robot_browser.js contains the default robot
credentials in plain text! This is only appropriate for
development on a local network behind a firewall.