first steps to dedicated server installation
I'm looking at resurrecting our old dedicated server installation
on an Amazon Lightsail instance. This will not be a hardened
server, but may help others explore remote teleoperation with
the Stretch RE1.

These two files are drafts for an installation script to be
run on a suitable Amazon Lightsail instance and a service
to start the coturn server when the instance is started.
They're in progress and documentation is forthcoming.