initial scripts to start & stop dedicated server
These are initial draft scripts to start and stop
a dedicated web server for remote teleoperation on an
Amazon Lightsail instance. They need work and
documentation. They will not result in a hardened
server. All of the dedicated server materials
in this repository are risky and should be used
with caution.