removed sudo from installation script instructions
I think sudo is unnecessary for the main installation script, since the script itself uses sudo when needed. I also suspect that some of the installation should be performed as the user. I'm taking the risky step of removing it from the instructions without testing the change. Hopefully, removing this now will cause less pain than leaving it in the documentation.