Tweak component visibility initialization
Add ability to leave a component hidden by default so that modal contents don't render until we show them later
Move settings.css into modal tree. When bootstrap shows a modal, it clones the modal element tree and puts it at the root of the document (in the light DOM). Putting the CSS link here ensures that it is copied and applied. This breaks the style encapsulation of the shadow DOM, but this is not something we're relying on at the moment anyways