Persist settings in model
Reorder settings component import so that the element is registered by the time it needs to be constructed in the operator page template
Instantiate settings component via the operator-page HTML template, so that the settings component can emit events and have them go into the operator page DOM tree
Emit "settingchanged" events when any form inputs change
In the operator-page, listen for settingchanged and persist the setting key-value pairs into local storage
When constructing operator page, configure settings modal form inputs to reflect the last saved settings
Add setting for control mode: incremental movement moves a joint a discrete step, continuous sends a velocity command and waits for it to be cancelled. Currently has no effect. The implementation of these control modes is in a separate commit
Rename component to settings-modal to match the pose-save-modal
Remove some dangling inline event handlers from settings modal template

These changes make it so that there is one "source of truth" for configuration; the model, which only knows how to persist key-value pairs, and obviate the need for the parent component to read the value of form inputs from within the settings component. The HTML contents of the component _are not part of its API_, only the settingchanged event is.