Distinguish discrete and continuous control commands
This commit adds the ability to send open-ended velocity commands to individual joints. If the "continuous" control mode is enabled in settings, clicking a region begins a velocity command, and clicking it again ends the velocity command.

Previous implementations of this feature relied on repeatedly sending discrete commands, but as documented in #30, this does not work well. Instead, we write trajectory commands that attempt to reach the joint limit in the appropriate direction with a desired velocity. This requires that we be able to cancel the currently executing velocity command, and that we consider how to prevent runaway commands should the operator lose their connection.

We modify the protocol that the operator side speaks to the robot to include straightforward "increment" and "velocity" movement commands. Velocity commands must be "affirmed" by the operator every 200ms or they are cancelled. The operator can either stop affirming, or issue an explicit stop command to end the velocity command.

To support switching between incremental and velocity commands, regions are now annotated with the joint that they control as well as the direction (positive or negative). This allows the click event handler to construct the appropriate call to the remoterobot, and to establish the affirmation heartbeat for the velocity commands.

To avoid complicating the direct control API that the remoterobot now exposes, the concept of "modifiers" for velocity commands has been pushed out of the robot layer all the way up to the operator page component. This means that the translation from the UI representation of velocity settings to actual joint commands (increment size or velocity) all happens within that component, and the robot can be indifferent to this complexity.

This implementation relies on an enhanced version of the follow_joint_trajectory shim from our simulation repository, hcrlab/hcrl_gazebo, that respects the passed in joint velocities as it rewrites the trajectory for the Gazebo ros_control trajectory server

Closes #30