Fix robot location display on interactive map
Add a new tfclient whose fixedFrame is map. Then use that to get the robot's location in map coordinates.
This is offset by the map offset on the operator side. Additionally after it is converted into pixels, the Y axis must be inverted.

Fix the operator side canvas drawing code to clear old robot positions